refactor: 合併 My Courses 與 Profile 頁面
