+++ client/src/App.jsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Profile from './components/Profile';
-import MyCourses from './pages/MyCourses';
 import PostCourse from './components/PostCourse';
 import CourseContent from './components/CourseContent';
 import AuthService from './services/auth.service';
@@ -17,7 +16,6 @@ export const ROUTER_PATH = {
   allCourses: '/all-courses',
   about: '/about',
   contact: '/contact',
-  myCourses: '/my-courses',
   profile: '/profile',
   postCourse: '/postCourse',
   courseContent: '/course-content',
@@ -41,8 +39,7 @@ function App() {
           <Route element={<AllCourses currentUser={currentUser} setIsModalOpen={setIsModalOpen} />} path={ROUTER_PATH.allCourses}></Route>
           <Route element={<About />} path={ROUTER_PATH.about}></Route>
           <Route element={<Contact />} path={ROUTER_PATH.contact}></Route>
-          <Route element={<Profile currentUser={currentUser} />} path={ROUTER_PATH.profile}></Route>
-          <Route element={<MyCourses currentUser={currentUser} setIsModalOpen={setIsModalOpen} />} path={ROUTER_PATH.myCourses}></Route>
+          <Route element={<Profile currentUser={currentUser} setIsModalOpen={setIsModalOpen} />} path={ROUTER_PATH.profile}></Route>
           <Route element={<PostCourse currentUser={currentUser} setIsModalOpen={setIsModalOpen} />} path={ROUTER_PATH.postCourse}></Route>
           <Route element={<CourseContent currentUser={currentUser} setIsModalOpen={setIsModalOpen} />} path={ROUTER_PATH.courseContent}></Route>
         </Route>
